% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clEval.R
\name{clInnerVariances}
\alias{clInnerVariances}
\title{Compute clusters' variances}
\usage{
clInnerVariances(data, clClassification, verbose = TRUE, ...)
}
\arguments{
\item{data}{Dataset of class type data.frame or matrix}

\item{clClassification}{Vector of samples' clustering labels}

\item{...}{Other arguments}
}
\value{
A list with:\cr
\itemize{
\item clVariances - data frame of variances nClustersxncol(data)
\item nClusters - number of clusters
\item sizeCl - number of elements by cluster
}
}
\description{
This function computes the inner variance for each cluster.
It takes in input a data.frame or a matrix of data and a vector
containing the classification of each sample.
}
