% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clEval.R
\name{clMeans}
\alias{clMeans}
\title{Compute centroids}
\usage{
clMeans(data, clClassification, verbose = TRUE, ...)
}
\arguments{
\item{data}{Dataset of class type data.frame or matrix}

\item{clClassification}{Vector of samples' clustering labels}

\item{...}{Other arguments}
}
\value{
A list with:\cr
\itemize{
\item clCentres - data frame of means nClustersxncol(data)
\item nClusters - number of clusters
\item sizeCl - number of elements by cluster
}
}
\description{
This function computes the centroids for each cluster.
It takes in input a data.frame or a matrix of data and a vector
containing the classification of each sample.
}
