% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clEval.R
\name{wcss}
\alias{wcss}
\title{Compute clusters' Within Cluster Sum of Squares}
\usage{
wcss(data, clClassification, verbose = TRUE, ...)
}
\arguments{
\item{data}{Dataset of class type data.frame or matrix}

\item{clClassification}{Vector of samples' clustering labels}

\item{...}{Other arguments}
}
\value{
A list with:\cr
\itemize{
\item WCSSByClByVar - WCSS by cluster by variable
\item WCSSByClB - WCSS by cluster
\item WCSS - Sum of the total WCSS
\item nClusters - number of clusters
\item sizeCl - number of elements by cluster
}
}
\description{
This function computes the Within Cluster Sum of Squares WCSS for each cluster.
It takes in input a data.frame or a matrix of data and a vector
containing the classification of each sample.
}
